+++ scrapy/spidermiddlewares/referer.py
@@ -11,14 +11,11 @@ from typing import TYPE_CHECKING, cast
 from urllib.parse import urlparse
 from warnings import warn
 
-from w3lib.url import safe_url_string
-
-from scrapy import Spider, signals
 from scrapy.exceptions import NotConfigured
 from scrapy.http import Request, Response
 from scrapy.spidermiddlewares.base import BaseSpiderMiddleware
 from scrapy.utils.misc import load_object
-from scrapy.utils.python import to_unicode
+from scrapy.utils.python import _looks_like_import_path, to_unicode
 from scrapy.utils.url import strip_url
 
 if TYPE_CHECKING:
@@ -288,56 +285,38 @@ class DefaultReferrerPolicy(NoReferrerWhenDowngradePolicy):
     name: str = POLICY_SCRAPY_DEFAULT
 
 
-_policy_classes: dict[str, type[ReferrerPolicy]] = {
-    p.name: p
-    for p in (
-        NoReferrerPolicy,
-        NoReferrerWhenDowngradePolicy,
-        SameOriginPolicy,
-        OriginPolicy,
-        StrictOriginPolicy,
-        OriginWhenCrossOriginPolicy,
-        StrictOriginWhenCrossOriginPolicy,
-        UnsafeUrlPolicy,
-        DefaultReferrerPolicy,
-    )
-}
-
-# Reference: https://www.w3.org/TR/referrer-policy/#referrer-policy-empty-string
-_policy_classes[""] = NoReferrerWhenDowngradePolicy
-
-
-def _load_policy_class(
-    policy: str, warning_only: bool = False
-) -> type[ReferrerPolicy] | None:
-    """
-    Expect a string for the path to the policy class,
-    otherwise try to interpret the string as a standard value
-    from https://www.w3.org/TR/referrer-policy/#referrer-policies
-    """
-    try:
-        return cast("type[ReferrerPolicy]", load_object(policy))
-    except ValueError:
-        tokens = [token.strip() for token in policy.lower().split(",")]
-        # https://www.w3.org/TR/referrer-policy/#parse-referrer-policy-from-header
-        for token in tokens[::-1]:
-            if token in _policy_classes:
-                return _policy_classes[token]
-
-        msg = f"Could not load referrer policy {policy!r}"
-        if not warning_only:
-            raise RuntimeError(msg)
-        warnings.warn(msg, RuntimeWarning)
-        return None
-
-
 class RefererMiddleware(BaseSpiderMiddleware):
     def __init__(self, settings: BaseSettings | None = None):  # pylint: disable=super-init-not-called
         self.default_policy: type[ReferrerPolicy] = DefaultReferrerPolicy
-        if settings is not None:
-            settings_policy = _load_policy_class(settings.get("REFERRER_POLICY"))
-            assert settings_policy
-            self.default_policy = settings_policy
+        self.policies: dict[str, type[ReferrerPolicy]] = {
+            p.name: p
+            for p in (
+                NoReferrerPolicy,
+                NoReferrerWhenDowngradePolicy,
+                SameOriginPolicy,
+                OriginPolicy,
+                StrictOriginPolicy,
+                OriginWhenCrossOriginPolicy,
+                StrictOriginWhenCrossOriginPolicy,
+                UnsafeUrlPolicy,
+                DefaultReferrerPolicy,
+            )
+        }
+        # Reference: https://www.w3.org/TR/referrer-policy/#referrer-policy-empty-string
+        self.policies[""] = NoReferrerWhenDowngradePolicy
+        if settings is None:
+            return
+        setting_policies = settings.getdict("REFERRER_POLICIES")
+        for policy_name, policy_class_import_path in setting_policies.items():
+            if policy_class_import_path is None:
+                del self.policies[policy_name]
+            else:
+                self.policies[policy_name] = load_object(policy_class_import_path)
+        settings_policy = self._load_policy_class(
+            settings.get("REFERRER_POLICY"), allow_import_path=True
+        )
+        assert settings_policy
+        self.default_policy = settings_policy
 
     @classmethod
     def from_crawler(cls, crawler: Crawler) -> Self:
@@ -381,16 +360,69 @@ class RefererMiddleware(BaseSpiderMiddleware):
                 DeprecationWarning,
                 stacklevel=2,
             )
+        allow_import_path = True
         policy_name = request.meta.get("referrer_policy")
         if policy_name is None and isinstance(response, Response):
             policy_header = response.headers.get("Referrer-Policy")
             if policy_header is not None:
                 policy_name = to_unicode(policy_header.decode("latin1"))
+                allow_import_path = False
         if policy_name is None:
             return self.default_policy()
-
-        cls = _load_policy_class(policy_name, warning_only=True)
+        cls = self._load_policy_class(
+            policy_name, warning_only=True, allow_import_path=allow_import_path
+        )
         return cls() if cls else self.default_policy()
+
+    def _load_policy_class(
+        self,
+        policy: str,
+        warning_only: bool = False,
+        *,
+        allow_import_path: bool = False,
+    ) -> type[ReferrerPolicy] | None:
+        """Load the :class:`ReferrerPolicy` class to use for *policy*.
+
+        *policy* may be any of the following:
+
+        -   A standard policy name, e.g. ``"no-referrer"``,
+            ``"origin-when-cross-origin"``, etc.
+
+        -   The special ``"scrapy-default"`` policy.
+
+        -   The import path of a :class:`ReferrerPolicy` subclass, e.g.
+            ``"scrapy.spidermiddlewares.referer.NoReferrerPolicy"`` or
+            ``"myproject.policies.CustomReferrerPolicy"``.
+
+        If *warning_only* is ``False`` (default) and *policy* cannot be turned
+        into a :class:`ReferrerPolicy` subclass, a :exc:`RuntimeError` is
+        raised. If *warning_only* is ``True``, a warning is logged and ``None``
+        is returned instead.
+
+        If *allow_import_path* is ``False`` (default), import paths are not
+        allowed, resulting in :exc:`RuntimeError` or ``None``. If ``True``,
+        they are allowed. Use ``True`` only if you trust the source of the
+        *policy* value.
+        """
+        if allow_import_path:
+            try:
+                return cast("type[ReferrerPolicy]", load_object(policy))
+            except ValueError:
+                pass
+        policy_names = [
+            policy_name.strip() for policy_name in policy.lower().split(",")
+        ]
+        # https://www.w3.org/TR/referrer-policy/#parse-referrer-policy-from-header
+        for policy_name in policy_names[::-1]:
+            if policy_name in self.policies:
+                return self.policies[policy_name]
+        msg = f"Could not load referrer policy {policy!r}"
+        if not allow_import_path and _looks_like_import_path(policy):
+            msg += " (import paths from the response Referrer-Policy header are not allowed)"
+        if not warning_only:
+            raise RuntimeError(msg)
+        warnings.warn(msg, RuntimeWarning)
+        return None
 
     def get_processed_request(
         self, request: Request, response: Response | None